requirements.txt
cloudpickle==1.3.0
cycler==0.10.0
graphviz==0.13.2
gym==0.17.1
gym-retro==0.7.1
kiwisolver==1.1.0
matplotlib==3.2.1
neat-python==0.92
numpy==1.18.2
opencv-python==4.2.0.32
pyglet==1.5.1
pyparsing==2.4.6
python-dateutil==2.8.1
scipy==1.4.1
six==1.14.0
